Reverse the order of satTyCxt and mkFun
Take for example `MaybeT (m :: * -> *) (a :: *)`.

Given `MaybeT :: (* -> *) -> * -> *`, we want to saturate the type
constructor with its type arguments to get something of kind `*`. The
type arguments are already in the type context `Σ` (a list, where the
first element is the last one added to the context):

    Σ = [∗, (∗ ⇒ ∗)]

The kind of `Maybe` is `(∗ ⇒ ∗) ⇒ ∗ ⇒ ∗`, now we want to saturate it to
`∗`. The result should be `MaybeT $ var (tl hd) $ var hd`. Previously,
`Σ` was in reverse order, as were the type variables. This was
confusing (certainly for tuples).

The arguments of a data constructor were also in reversed order, this
was fixed by reversing the order in `mkFun`.