Change the encoding of ADTs
Encode all information (type constructors, type arguments, data
constructors, foreign names, ...) of an Algebraic Data Type in a single
record. The corresponding type constructor can be obtained by wrapping
the record in `con`. The data constructors can be obtained similarly,
but the index of the data constructor must be passed as well.

The main advantage is that, given either a type or a data constructor,
all others can easily be reconstructed by reusing the wrapped ADT. E.g.,
given the type constructor, we can enumerate all data constructors,
which is necessary for checking exhaustiveness of pattern matching and
for generating all cases when deriving code for an ADT. Inversely, given
a data constructor, the type constructor can easily be reconstructed,
which is necessary when type checking.